Create Url Shortener Project - fixed update model with attributes count_of_use and add access rule for simple user;
TODO make api

+++ controllers/UserController.php
@@ -4,6 +4,8 @@ namespace app\controllers;
 
 
 use app\models\search\UserSearch;
+use app\objects\CheckUserAccess;
+use function var_dump;
 use Yii;
 use app\models\User;
 
@@ -25,7 +27,7 @@ class UserController extends Controller
             'verbs' => [
                 'class' => VerbFilter::className(),
                 'actions' => [
-                    'delete' => ['POST'],
+                    'delete' => ['POST', 'GET', 'PUT'],
                 ],
             ],
         ];
@@ -40,16 +42,14 @@ class UserController extends Controller
         $searchModel = new UserSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        if (Yii::$app->user->id === User::find()->where(['username' => 'mrnkravchenko@gmail.com'])->one()->id) {
+        if (CheckUserAccess::isAdmin()){
 
             return $this->render('index', [
                 'searchModel' => $searchModel,
                 'dataProvider' => $dataProvider,
             ]);
 
-
         } else return $this->goHome();
-
 
     }
 
@@ -61,10 +61,14 @@ class UserController extends Controller
      */
     public function actionView($id)
     {
-//        TODO Доделать - закрыть доступ к просмотру, обновлению и редактированию пользователей, кроме Админа
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if (CheckUserAccess::isAdmin()){
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+
+        } else return $this->goHome();
+
+
     }
 
     /**
@@ -117,25 +121,38 @@ class UserController extends Controller
     {
         $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
-        }
+        if ($model->id === Yii::$app->user->id || CheckUserAccess::isAdmin()) {
 
-        return $this->render('update', [
-            'model' => $model,
-        ]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+            ]);
+
+        } else return $this->goHome();
+
+
     }
 
     /**
      * Deletes an existing User model.
      * If deletion is successful, the browser will be redirected to the 'index' page.
+     *
      * @param integer $id
+     *
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
+     * @throws \Throwable
+     * @throws \yii\db\StaleObjectException
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if (CheckUserAccess::isAdmin()){
+            $this->findModel($id)->delete();
+        }
+
 
         return $this->redirect(['index']);
     }
